Move Adapters para novo pacote "adapter"
